+++ src/components/article/zoom-in-leading-image.js
@@ -1,5 +1,5 @@
 import FadeText from '@twreporter/react-components/lib/fade-text'
-import Header from '@twreporter/react-components/lib/header'
+import HeaderComposite from '@twreporter/react-components/lib/header'
 import PropTypes from 'prop-types'
 import React from 'react'
 import TitleRowUpon from './title-row-upon'
@@ -12,6 +12,8 @@ import { SITE_META } from '../../constants/index'
 import { camelizeKeys } from 'humps'
 import { screen } from '../../themes/screen'
 
+
+const { Header } = HeaderComposite
 const { childAnimationStoper, unlockAfterAnimation } = FadeText.scrollManager
 const topNavbarHeight = 151