+++ src/containers/Article.js
@@ -7,6 +7,7 @@ import ArticleTools from './ArticleTools'
 import Header from '@twreporter/react-components/lib/header'
 import Helmet from 'react-helmet'
 import LeadingVideo from '../components/shared/LeadingVideo'
+import License from '../components/shared/License'
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import ReadingProgress from '../components/article/ReadingProgress'
@@ -400,6 +401,9 @@ class Article extends PureComponent {
       titleFontColor: titleColor,
       subtitleFontColor: subTitleColor
     }
+
+    const license = _.get(article, 'copyright', 'Creative-Commons')
+
     return (
       <div>
         <Helmet
@@ -530,6 +534,7 @@ class Article extends PureComponent {
                 />
               </article>
             </Content>
+            <License license={license} publishedDate={article.publishedDate}/>
             <div ref={div => {this.progressEnding = div}}
                 className={cx(commonStyles['components'], 'hidden-print', styles['padding-patch'])}>
               <div className={cx('inner-max', commonStyles['component'])}>